Add pkg audit check of new version
Instead of only the current version, also add check the new version
for known vulnerabilities. Use ! as flag indicator.
There is no distinction between:
- the new version has the same issue(s) as the current version
- the new version has less issue(s) than the current version, but
  still has some
- the current version is fine, and the new version introduces the
  problem

This is intended to make it clear that portmaster needs to be handed
the `-m DISABLE_VULNERABILITIES=YES` option.

Yes, this make option should be called DISABLE_VULNERABILITY_CHECK.

Update the readme to reflect the new ! flag.