Fix false-positive ! flagging
* The package name rewriting by bsd.default-versions.mk
could lead to `pkg audit` returning a false positive
result when checking `pkg audit $name-$nversion` where
$name is the name as returned by `pkg query %n $origin`
and $nversion is the version of the port available for
installation.
For this to happen, the a port must exist in multiple
versions (foo/bar10 and foo/bar11), the ports default
being foo/bar11 and DEFAULT_VERSIONS= bar=1.1 being set.
That way foo/bar11 will be registered as bar instead of
bar10 and `pkg audit` will compare the 1.0 version number
against the 1.1 version number of a different port.

* remove note about this from README.md